[fix]hilt 의존성 추가 & 리포트 데이터 수정 (#108)

+++ app/src/main/java/com/example/pauze/data/model/ReportModel.kt
@@ -9,7 +9,7 @@ data class ChartBar(val label: String, val score: Int)
 data class WeeklyReportDto(
     val averageScore: Double,
     val hardestDay: String,
-    val hardestScore: Long,
+    val hardestScore: Double,
     val pauzeCount: Long,
     val scoreChange: Double?,
     val dailyScores: List<DailyScore>,
@@ -31,7 +31,7 @@ data class TopTrigger(
 data class MonthlyReportDto(
     val averageScore: Double,
     val hardestWeek: String,
-    val hardestScore: Long,
+    val hardestScore: Double,
     val pauzeCount: Long,
     val scoreChange: Double?,
     val weeklyScores: List<WeeklyScore>,
